refactor(app:ccf-organ-info): :recycle: replaced old component with new one
