--- apps/ccf-organ-info-new/eslint.config.mjs
@@ -1,36 +1,0 @@
-import { configs } from '../../eslint.config.mjs';
-
-export default [
-  ...configs.base,
-  ...configs.angular,
-  {
-    files: ['**/*.ts'],
-    rules: {
-      '@angular-eslint/directive-selector': [
-        'error',
-        {
-          type: 'attribute',
-          prefix: 'ccf',
-          style: 'camelCase',
-        },
-      ],
-      '@angular-eslint/component-selector': [
-        'error',
-        {
-          type: 'element',
-          prefix: 'ccf',
-          style: 'kebab-case',
-        },
-      ],
-      '@angular-eslint/prefer-signals': 'off',
-      '@angular-eslint/prefer-standalone': 'off',
-    },
-  },
-  {
-    files: ['**/*.html'],
-    rules: {
-      '@angular-eslint/template/click-events-have-key-events': 'warn',
-      '@angular-eslint/template/interactive-supports-focus': 'warn',
-    },
-  },
-];
