--- apps/ccf-organ-info-new/project.json
@@ -1,70 +1,0 @@
-{
-  "name": "ccf-organ-info-new",
-  "metadata": {
-    "version": "4.0.0"
-  },
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "application",
-  "sourceRoot": "apps/ccf-organ-info-new/src",
-  "prefix": "app",
-  "tags": ["type:app", "project:ccf"],
-  "targets": {
-    "build": {
-      "executor": "@nx/angular:application",
-      "options": {
-        "assets": [
-          "apps/ccf-organ-info-new/src/favicon.png",
-          "apps/ccf-organ-info-new/src/assets",
-          {
-            "input": "libs/design-system/assets",
-            "glob": "**/*",
-            "output": "./assets"
-          }
-        ],
-        "polyfills": ["apps/ccf-organ-info-new/src/polyfills.ts"]
-      },
-      "configurations": {
-        "production": {
-          "fileReplacements": [
-            {
-              "replace": "apps/ccf-organ-info-new/src/environments/environment.ts",
-              "with": "apps/ccf-organ-info-new/src/environments/environment.prod.ts"
-            }
-          ],
-          "budgets": [
-            {
-              "type": "initial",
-              "maximumWarning": "2mb",
-              "maximumError": "5mb"
-            },
-            {
-              "type": "anyComponentStyle",
-              "maximumWarning": "6kb",
-              "maximumError": "10kb"
-            }
-          ]
-        },
-        "staging": {
-          "fileReplacements": [
-            {
-              "replace": "apps/ccf-organ-info-new/src/environments/environment.ts",
-              "with": "apps/ccf-organ-info-new/src/environments/environment.staging.ts"
-            }
-          ]
-        }
-      }
-    },
-    "serve": {
-      "executor": "@nx/angular:dev-server"
-    },
-    "compodoc": {
-      "executor": "@twittwer/compodoc:compodoc",
-      "options": {
-        "tsConfig": "apps/ccf-organ-info-new/tsconfig.app.json"
-      }
-    },
-    "build-webcomponent": {
-      "command": "node tools/scripts/bundle-scripts.mjs dist/apps/ccf-organ-info-new/browser/"
-    }
-  }
-}
