--- apps/ccf-organ-info-new/src/app/app.component.html
@@ -1,77 +1,0 @@
-<div class="header">
-  <div class="left">
-    <div class="icons">
-      <hra-icon [svgIcon]="'product:eui'"></hra-icon>
-      <hra-icon [svgIcon]="getOrganLogo()"></hra-icon>
-    </div>
-    <div class="title">Organ Information</div>
-  </div>
-  <button mat-icon-button hraIconButtonSize="large" hraIconButtonVariant="dark" [matMenuTriggerFor]="headerMenu">
-    <mat-icon>more_vert</mat-icon>
-  </button>
-</div>
-<div class="content">
-  <div class="info-view">
-    <div class="organ-title">
-      {{ statsLabel$ | async }}
-    </div>
-    <div class="params">
-      <mat-button-toggle-group
-        *ngIf="sex$ | async"
-        hraButtonToggleSize="medium"
-        disabled="false"
-        [hideSingleSelectionIndicator]="false"
-        [value]="sex$ | async"
-        (change)="updateInput('sex', $event.value)"
-      >
-        <mat-button-toggle value="Male">Male</mat-button-toggle>
-        <mat-button-toggle value="Female">Female</mat-button-toggle>
-      </mat-button-toggle-group>
-      <mat-button-toggle-group
-        *ngIf="side$ | async"
-        hraButtonToggleSize="medium"
-        disabled="false"
-        [hideSingleSelectionIndicator]="false"
-        [value]="side$ | async"
-        (change)="updateInput('side', $event.value)"
-      >
-        <mat-button-toggle value="Left">Left</mat-button-toggle>
-        <mat-button-toggle value="Right">Right</mat-button-toggle>
-      </mat-button-toggle-group>
-    </div>
-    <div class="stats-table">
-      <hra-table
-        [columns]="tableColumns"
-        [rows]="(stats$ | async) ?? []"
-        [style]="'alternating'"
-        [verticalDividers]="true"
-      ></hra-table>
-    </div>
-  </div>
-  <div class="organ-view">
-    <ccf-organ
-      [blocks]="(blocks$ | async) ?? []"
-      [filter]="$any(filter$ | async)"
-      [sex]="(sex$ | async)!"
-      [side]="$any(side$ | async)"
-      [organ]="(organ$ | async) ?? undefined"
-      [scene]="(scene$ | async) ?? []"
-      (nodeClick)="nodeClicked.emit($event)"
-    >
-    </ccf-organ>
-  </div>
-
-  <mat-menu #headerMenu="matMenu">
-    <a mat-menu-item [href]="(organ$ | async)?.['@id']" target="_blank">
-      <mat-icon>database</mat-icon>
-      Source data
-    </a>
-    <a mat-menu-item href="https://docs.humanatlas.io/dev/web-components#eui-organ-information" target="_blank">
-      <mat-icon>help</mat-icon>
-      Help & documentation
-    </a>
-    <mat-divider class="menu-divider"></mat-divider>
-    <a mat-menu-item href="https://humanatlas.io/publications" target="_blank">View publications</a>
-    <a mat-menu-item href="https://humanatlas.io" target="_blank">Human Reference Atlas Portal</a>
-  </mat-menu>
-</div>
