--- apps/ccf-organ-info-new/src/app/app.component.spec.ts
@@ -1,5 +1,0 @@
-describe('AppComponent', () => {
-  it('should pass the placeholder test', () => {
-    expect(true).toBeTruthy();
-  });
-});
