--- apps/ccf-organ-info-new/src/index.html
@@ -1,19 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>CCF Organ Info</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/png" href="favicon.png" />
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500&display=swap" rel="stylesheet" />
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons|Material+Icons+Outlined" rel="stylesheet" />
-  </head>
-
-  <body class="mat-typography">
-    <ccf-organ-info
-      organ-iri="http://purl.obolibrary.org/obo/UBERON_0002113"
-      highlight-providers="TMC-Vanderbilt"
-      data-sources='["https://purl.humanatlas.io/collection/ds-graphs"]'
-    ></ccf-organ-info>
-  </body>
-</html>
