--- apps/ccf-organ-info-new/src/styles.scss
@@ -1,11 +1,0 @@
-@use 'light-theme';
-@use 'theming' as design-system-theming;
-
-html {
-  @include design-system-theming.theme(light-theme.$theme);
-}
-
-html,
-body {
-  height: 100%;
-}
