--- apps/ccf-organ-info-new/src/webcomponent-example.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>CCF Organ Info</title>
-    <base href="./" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="icon" type="image/x-icon" href="favicon.ico" />
-    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500&amp;display=swap" rel="stylesheet" />
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons|Material+Icons+Outlined" rel="stylesheet" />
-    <link href="styles.css" rel="stylesheet" />
-    <script src="wc.js" async></script>
-  </head>
-  <body class="mat-typography">
-    <!-- <ccf-organ-info
-    organ-iri="http://purl.obolibrary.org/obo/UBERON_0004538"
-    hubmap-data-service="search-api"
-    hubmap-portal-url="https://portal.test.hubmapconsortium.org/"
-    hubmap-data-url="https://search.api.hubmapconsortium.org/v3/entities/search"
-    hubmap-asset-url="https://assets.test.hubmapconsortium.org"
-  ></ccf-organ-info> -->
-    <ccf-organ-info organ-iri="http://purl.obolibrary.org/obo/UBERON_0002113"></ccf-organ-info>
-  </body>
-</html>
