--- apps/ccf-organ-info-new/tsconfig.app.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": []
-  },
-  "files": ["src/main.ts", "src/polyfills.ts"],
-  "include": ["src/**/*.ts"],
-  "exclude": ["jest.config.ts", "src/test-setup.ts", "**/*.spec.ts"]
-}
