--- apps/ccf-organ-info-new/tsconfig.editor.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "include": ["**/*.ts"],
-  "compilerOptions": {},
-  "exclude": ["jest.config.ts", "src/test-setup.ts", "**/*.spec.ts"]
-}
