--- apps/ccf-organ-info-new/tsconfig.spec.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "types": ["jest", "node", "@testing-library/jest-dom"]
-  },
-  "files": ["src/test-setup.ts"],
-  "include": ["jest.config.ts", "**/*.spec.ts", "**/*.d.ts"]
-}
