+++ apps/ccf-organ-info/src/app/components/organ/organ.module.ts
