--- apps/ccf-organ-info/src/app/features/organ/organ.component.html
@@ -1,30 +1,0 @@
-<ccf-body-ui
-  #bodyUI
-  class="body-ui"
-  [scene]="scene"
-  [interactive]="true"
-  camera="perspective"
-  (nodeClick)="nodeClicked($event)"
-  (initialized)="zoomToFitOrgan()"
->
-</ccf-body-ui>
-
-<div class="stage-options">
-  <ccf-slide-toggle
-    class="view-slider"
-    [labels]="['Male', 'Female']"
-    [value]="sex!"
-    [hidden]="!sex"
-    (valueChange)="updateSex($any($event))"
-  >
-  </ccf-slide-toggle>
-
-  <ccf-slide-toggle
-    class="view-slider"
-    [labels]="['Left', 'Right']"
-    [value]="side!"
-    [hidden]="!side"
-    (valueChange)="updateSide($any($event))"
-  >
-  </ccf-slide-toggle>
-</div>
