--- apps/ccf-organ-info/src/app/features/slide-toggle/slide-toggle/slide-toggle.module.ts
@@ -1,12 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-
-import { SlideToggleComponent } from './slide-toggle.component';
-
-@NgModule({
-  declarations: [SlideToggleComponent],
-  imports: [CommonModule, MatSlideToggleModule],
-  exports: [SlideToggleComponent],
-})
-export class SlideToggleModule {}
