--- apps/ccf-organ-info/src/app/modules/stats-list/stats-list.component.scss
@@ -1,25 +1,0 @@
-.stats-container {
-  margin-bottom: 2rem;
-  min-height: 9rem;
-
-  .header {
-    margin-bottom: 1rem;
-  }
-
-  .stats-row {
-    display: flex;
-    flex-direction: row;
-    height: 1.325rem;
-
-    .stat {
-      width: 2rem;
-      margin-right: 1rem;
-      text-align: right;
-    }
-
-    .label {
-      flex-grow: 1;
-      text-align: left;
-    }
-  }
-}
