--- apps/ccf-organ-info/src/app/modules/stats-list/stats-list.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { StatsListComponent } from './stats-list.component';
-
-@NgModule({
-  declarations: [StatsListComponent],
-  imports: [CommonModule],
-  exports: [StatsListComponent],
-})
-export class StatsListModule {}
