+++ apps/ccf-organ-info/src/webcomponent-example.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>HuBMAP CCF Organ Info</title>
+    <title>CCF Organ Info</title>
     <base href="./" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <link rel="icon" type="image/x-icon" href="favicon.ico" />
